renamed a scout lib
