+++ scout/commands/load/exons.py
@@ -9,7 +9,7 @@ from flask.cli import with_appcontext
 from scout.load import load_exons
 
 from scout.utils.handle import get_file_handle
-from scout.utils.requests import fetch_ensembl_exons
+from scout.utils.scout_requests import fetch_ensembl_exons
 from scout.server.extensions import store
 
 LOG = logging.getLogger(__name__)
