+++ scout/load/transcript.py
@@ -4,7 +4,7 @@ from pprint import pprint as pp
 
 from click import progressbar
 
-from scout.utils.requests import fetch_ensembl_transcripts
+from scout.utils.scout_requests import fetch_ensembl_transcripts
 from scout.parse.ensembl import parse_transcripts
 from scout.build.genes.transcript import build_transcript
